extract outro
